tests-qualidade-soft/activities/server/tests/integration/container.spec.ts
import createContainer from "../../src/container";

describe('Container', () => {
    let container;

    test('It needs to create a mongodb client', async () => {
        container = await createContainer();
        const client = await container.getClient();
        expect(client).not.toBe(null);
        expect(client).not.toBe(undefined);
    });

    test('It needs to return always the same instance', async () => {
        const clientA = await container.getClient();
        const clientB = await container.getClient();
        expect(clientA).toStrictEqual(clientB);
    });

    test('It needs to create a repository of events', async () => {
        const repository = await container.getRepository();
        expect(repository).not.toBe(null);
        expect(repository).not.toBe(undefined);
    });

});
